pyproject.toml
[project]
name = "ai-procrastination-detector"
version = "1.0.0"
description = "An intelligent desktop application that detects cell phone usage via webcam and triggers humorous alerts to help you stay focused"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["productivity", "ai", "yolo", "computer-vision", "procrastination", "focus"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video :: Capture",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "torchaudio>=2.5.1",
    "ultralytics>=8.3.40",
    "opencv-python>=4.10.0.84",
    "pygame>=2.6.1",
    "pyyaml>=6.0.2",
    "imageio>=2.34.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/hourse-of-wisdom"
Repository = "https://github.com/YOUR_USERNAME/hourse-of-wisdom"
Issues = "https://github.com/YOUR_USERNAME/hourse-of-wisdom/issues"

[project.scripts]
procrastination-detector = "app:main"

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch" }]
torchvision = [{ index = "pytorch" }]
torchaudio = [{ index = "pytorch" }]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
